code review feedback: move getScrollContainer to dom, don't handle outside viewport
